+++ changelog.md
@@ -1,3 +1,9 @@
+## 0.0.3 (2025-01-24)
+- Don't submit multi-selects as comma-separated values. Use hidden select instead.
+
+## 0.0.2 (2025-01-22)
+- Allow customizing the base template
+
 ## 0.0.1 (2024-11-28)
 
 - Use html instead of .createElement to make the code more readable